examples/ai-graph-generator/requirements.txt
streamlit>=1.30.0
openai>=1.0.0
python-dotenv>=1.0.0
e2b>=0.13.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
pandas>=2.1.0
numpy>=1.24.0
asyncio>=3.4.3